lambdas + overloaded specs in gradual mode: implementation
Summary:
Keeping the previous logic in strict mode, but relaxing in gradual mode.
The trade-off in gradual mode is the same as previously: when a single sub-spec of the overloaded spec cannot be selected, we approximate with `dynamic()` - this may result in accepting erroneous code in gradual mode.
(The trade-off would be mitigated once we add more signal about approximations in gradual mode).

Reviewed By: VLanvin, perehonchuk, TD5

Differential Revision: D38700863

fbshipit-source-id: f03aecdef232d4be67517d55f6141d8f79edb951