Better refinements of dynamic() w/ gradual subtyping
Summary:
Better handling of `dynamic()` in subtyping, occurrence typing, and constraint solving.

### Subtyping changes

Changes to subtyping come from the theory developed in [1] (mostly chap. 6, p. 122-125), and are summarised below.

- Modified subtyping to memorize the polarity of both types at hand. This polarity info is used when reaching an instance of the dynamic type. Basically, positive instances of dynamic are considered to be `term()`, while negative instances are considered to be `none()`. The previous behaviour of subtyping is obtained by checking whether `neg(t1) < pos(t2)`.
Since `neg(dynamic()) = none()` and `pos(dynamic()) = term()`, this gives a relation where `dynamic()` is both a subtype and supertype of everything.

- Introduced gradual subtyping, which checks whether `neg(t1) < neg(t2)` *and* `pos(t1) < pos(t2)`. For this relation, `dynamic()` is only a subtype of itself. This allows better handling of unions, since for example, `dynamic() | number()` is not equivalent to `dynamic()` for gradual subtyping.

### Constraint solving changes

The change to constraint solving is simple: whenever a variable `X` occurs in covariant position and its only constraint is `none() < X`, then we infer `dynamic()` instead of `none()`. This is still sound since the result of constraint solving is checked with normal subtyping (for which `dynamic() < none()`), and allows for better handling of dynamic lambdas.

[1] Lanvin, Victor. [A Semantic Foundation for Gradual Set-Theoretic Types.](https://vlanvin.fr/papers/thesis.pdf) Diss. PhD thesis, Université de Paris, 2021.

Reviewed By: ilya-klyuchnikov

Differential Revision: D38824968

fbshipit-source-id: cc21e4adce46be3765c54765509d4198a09f59f7