better case<->if understanding: implementation
Summary:
In big code bases there are patterns like

```
case is_number(X) and other_cond(...) of
  % we may use X as number here
  true -> ..
  false -> ..
end
```

The current implementation generalizes previous handling of such cases in a simple way - by rewriting it into an `if-expression` (a selection should be guard-compatible in this case)

Upside: simplicity.
Downside: conditions in selector may go beyond guards - in this case we do nothing so far.

Reviewed By: VLanvin

Differential Revision: D38895703

fbshipit-source-id: 5230fba480bd18d6ff2a8b5f6785bff930d40936