Algorithmic optimizations to occurrence typing
Summary:
Features two main changes to occurrence typing to better support large number of clauses and large unions.

### 1. Collection of relevant negative propositions only

When computing the list of props that are used to refine the types in a clause, `combine` uses the positive type information of the context to filter useless negative propositions inherited from previous clauses.
The main logic is that `Neg(obj, t)` can be ignored if `Pos(obj, t2)` is present in the context, and `t` and `t2` do not overlap.

### 2. Immediate application of props to environments

Rather than computing the DNF on props and then applying this DNF on the current type environment, the two operations are done simultaneously by immediately applying the props on the current type environment.
This allows to incrementally filter out redundant environments, thus avoiding an exponential blowup of the number of environments at hand.
This is done by calling `keepBestEnvs` (which relies on `envSubtype`) during the computation of the DNF.

Reviewed By: ilya-klyuchnikov

Differential Revision: D39507209

fbshipit-source-id: 5e6335647cba472bc28420926a40355f05026c48