Support for variable elimination in non-covariant aliases
Summary:
Improve variable elimination to support non-covariant aliases.
Contravariance is handled by simply switching the variance, as for function types.
Invariance is handled similarly to Typed Racket: eliminating a variable X in a type T returns Any in Promote mode, and None in Demote mode. This is not the most general solution since all the information in T is lost, but a sound one nonetheless.

Reviewed By: ilya-klyuchnikov

Differential Revision: D41995158

fbshipit-source-id: 833e9e6e776244207f14709afc29c12547cb06e7