Better solving of union constraints
Summary:
Union constraints `T <: T1 | T2` are solved by eliminating variables in all types, and finding the best candidate (between `T1` and `T2`) to solve.

This was only taking place when the right hand side contained variables, however, due to contravariance, variables may appear in the left hand side instead.

This diff instead checks early that one of the bounds contains variables, and aborts constraint solving if it's not the case.

Reviewed By: ilya-klyuchnikov

Differential Revision: D42415423

fbshipit-source-id: fc65feca25610cef0dccb4a8cb8ddc8d5c233104