More precise inference with dynamic()
Summary:
Make unification of `dynamic()` with variables under constructors more precise.
In essence, when solving `dynamic() < type()`, all free variables in `type()` are now constrained between `dynamic()` and `term()`.

Similar reasoning may be needed when solving `type() < dynamic()` due to contravariance, but postponing it for now as I'm not sure of the implications.

Reviewed By: ilya-klyuchnikov

Differential Revision: D43270377

fbshipit-source-id: dc3193c96800d7019e6f2f59a81b154ea32ecfb2