Introduce limit for refinement of union type aliases
Summary:
Do not refine union type aliases in occurrence typing if they are above a certain size.

As with similar limits, it can be bypassed by using the pragma `-unlimited_refinement` on the corresponding function.

## Explanation

Oftentimes, such large union aliases are simply enumeration of atoms, and are not meant to be checked exhaustively. For example, the following pattern occurs frequently:
```
-type large_union() :: a | ...

-spec handle_a(large_union()) -> ok | {error, large_union()}.
handle_a(a) -> ok;
handle_a(V) -> {error, V}.
```
Refining the type of `V` here means removing `a` from `large_union()`, hence creating a large unnamed union type. Then, to type-check the second branch, eqWAlizer needs to ensure that each component of this unnamed union type is a subtype of `large_union()`, which induces a O(n^2) subtyping check where `n` is the size of `large_union()`.

By keeping the type of `V` as `large_union()`, the subtyping check is O(1).

Reviewed By: ilya-klyuchnikov

Differential Revision: D43497931

fbshipit-source-id: a73e9e992a4c64bc0a29bc3aa3300e6f9927c670