Always use elab mode for records
Summary:
Without this diff, record field access uses elab mode only if the field is explicitly tagged as refinable. This can lead to confusing signal, especially if the field is expected to be refined implicitly by occurrence typing/narrowing. This also introduces an inconsistency between record patterns and guards with field selection.

This diff opts to rely on elaboration entirely for field selection, as the tradeoff in signal quality seems worth it.

Reviewed By: ilya-klyuchnikov

Differential Revision: D44089169

fbshipit-source-id: b66d3dfc6a5eb7548978d122ed1ebd7a6e624cdb