Deeper refinements with guards
Summary:
Many type refinements only take place when the tested expression is a variable.
This diff refactors Occurrence so that support for more complex test expressions can be added.
It also adds support for test expressions that use `hd/1`, such as `hd(L) =:= a`, as it is a fairly common pattern.

Reviewed By: ilya-klyuchnikov

Differential Revision: D44054959

fbshipit-source-id: 44b466aac86caff67905d003117fd787267590f8