Serialization of eqWAlizer AST
Summary:
Add serialization/deserialization of converted AST.

Serialization is handled out-of-the-box by serde on ELP's end. The only particularity is that BigInts are wrapped into a transparent struct to be serialized as strings.

Deserialization is handled by jsoniter-scala on eqWAlizer's end. Differences in case and naming are handled on this end. BigInts are tagged using `stringified`.

Reviewed By: RobinMorisset

Differential Revision: D44688327

fbshipit-source-id: 27c46885556fcb73b980cd4cac12e4789e166cbd