Make IPC protocol easier to extend
Summary:
The IPC protocol is getting more and more complex the more we migrate stubs in ELP.

This simplifies it by removing `GetStubsBytes` and instead relying only on `GetAstBytes`, which is now parameterised by an enumeration which can be easily extended with more cases (e.g. `ExtendedStubs`).

This is compatible with `mini-elp` out of the box, since `mini-elp` treats `GetStubsBytes` and `GetAstBytes` equally, and ignores additional fields in the request.

Reviewed By: ilya-klyuchnikov, michalmuskala

Differential Revision: D46147228

fbshipit-source-id: 3fee797f44a6bd0ea4094da96d47cf02cd07e516