Custom serialisation/deserialisation of ids
Summary:
Use a custom serialiser/deserialiser for ids in ELP and eqWAlizer. Ids are now serialized as a string `id/arity`.

On ELP's end, this uses `serde_with` to derive a serialiser from the `Display` trait.

This is needed to properly serialise maps where ids are used as keys, as in `ModuleStub.scala` for example. For now, remote ids are still serialised as before.

Reviewed By: ilya-klyuchnikov, michalmuskala

Differential Revision: D46147226

fbshipit-source-id: 8223fec49ff43dc9201d08f472cd9f3c207dc007