Fixes to dependency tracking for ELP shell
Summary:
Dependencies introduced during the TransValid step of stubs validation were not properly tracked on ELP's end, leading to some broken state in ELP shell.

This diff introduces a temporary fix by tracking dependencies introduced during the validation of each stub, and batches them with the dependencies sent to ELP at the end of the eqWAlization.

This will be removed as soon as the TransValid phase is moved to ELP.

Reviewed By: ilya-klyuchnikov

Differential Revision: D46897035

fbshipit-source-id: 75b2de09eae1f146f71d4e6ea3cac65b7d462a94