Support for maybe
Summary:
Support for `maybe` expressions in eqWAlizer. A few notes:

- For clarity purposes, this distinguishes between `Maybe` (without else clauses) and `MaybeElse` (with else clauses).
- `MaybeMatch` are added as top-level expressions following the same typing rules as `Match`.
- This is a first, best-effort attempt for quick support. There is no fancy features such as occurrence typing for `else` clauses, and the type of `MaybeMatch` expressions is not tracked. `else` clauses are typed using `dynamic()` for the type of the argument in gradual mode, and `term()` in static mode.

### Additional notes about `Maybe`

`Maybe` is typed by elaborating the type of expressions appearing in the RHS of a `MaybeMatch`. In check mode, these types are checked individually against the expected type to get more precise signal.

The type matched by the pattern is not subtracted from the type of the RHS, for the sake of simplicity. This means that
```
-spec foo(a | b) -> a.
foo(V) ->
  maybe
    b ?= V,
    a
  end.
```
will not type-check, as `b ?= V` will be assumed to shortcut the maybe with type `a | b`.

Reviewed By: ilya-klyuchnikov

Differential Revision: D47840288

fbshipit-source-id: 9d4edc4c11ae4dc762e466a23b9522cb93b970de