Collect and send type info
Summary:
Add a new module `TypeInfo` storing a global hashmap `module -> pos -> type` to store type information during type-checking.

Type info in a given range can be erased using `clear(pos: Pos)` to accommodate for error tolerance.

Type information for variables is added to this table during type-checking, and is then sent to ELP along with the diagnostics. This is only performed in IDE mode, to avoid unnecessary computations.

Reviewed By: ilya-klyuchnikov

Differential Revision: D49686985

fbshipit-source-id: c5efe62d8f6112b431234c2ed4a870cf60d9a37e