Use new IPC protocol by default (except for mini-elp)
Summary:
A new IPC protocol was introduced for ELP shell, which allows for precise by-module caching of dependencies even when running `eqwalize-all`.
The previous protocol was kept around to avoid regressions in IDE and in ELP CLI.

After a period of testing, this protocol seems stable enough that the previous one can be deprecated, except for mini-elp.

Reviewed By: ilya-klyuchnikov, alanz

Differential Revision: D51524775

fbshipit-source-id: 9de5ff1cb4e2e9faf40d634917b8b171df0c4e80