Temporary bandaid for incomplete transitive check
Summary:
Change an unchecked `.get` (which should have been sound) to return `DynamicType` if it fails.
This is not ideal, and temporary while the transitive check algorithm is fixed.

This should not actually allow type errors to slip through, since it can only occur if there is an invalid type somewhere, that will be accurately reported anyway.

Reviewed By: ilya-klyuchnikov

Differential Revision: D51938905

fbshipit-source-id: 30e22793e4d6a5454a1c00e809ef78506bfc0917