Improvements to maps:fold with shapes
Summary:
Improves inference when the lambda acts on shapes.

The main logic is the function `narrow.joinAndMergeShapes` which joins a set of types using `subtype.join`, but "merges" shapes so that:
`#S{a => a}` joined with `#S{b => b}` is `#S{a => a, b => b}`
`#S{a => a}` joined with `#S{a => b}` is `#S{a => a | b}`
`#S{a := a}` joined with `#S{a := a, b => b}` is `#S{a := a, b => b}`

This function is called from `maps:fold` elaboration, only at the first elaboration step. Then the merged type is used as the accumulator type for the second elaboration step, whose result is merged with `subtype.join` to avoid merging shapes that should not be.

Reviewed By: ilya-klyuchnikov

Differential Revision: D54900007

fbshipit-source-id: 619ad636b4e8ebd2239fe2183b04db3d9a034c23