Fix subtyping dict < shape being too lenient
Summary:
The root cause of D56416941 is subtyping being too lenient for dynamic dicts to shapes, causing dynamic dicts to be "approximated away" when joining types.

This fixes it by considering `#{dynamic() => dynamic()}` to be a subtype of any shape only when the subtyping polarity of the dict is negative.

Also change `narrow.adjustMapType` to preserve dynamic maps, to avoid introducing too much noise.

Reviewed By: ilya-klyuchnikov

Differential Revision: D56416893

fbshipit-source-id: 88e3f68b9e88f6d8b61556a3edbbdad2933a4235