CustomReturn for erlang:min/2
Summary:
It turns out that calls `erlang:min(X, Y)` when `X` is a number and `Y` is `number() | infinity` or `number() | undefined` are not so uncommon - and can be considered as an "Erlangisms".  Programmers rely on the fact that `min(1, infinity) -> 1` or `min(1, undefined) -> 1`.

Very ad-hoc custom handling for such cases.

Reviewed By: VLanvin

Differential Revision: D57249552

fbshipit-source-id: 6bedce5381d23d60b7d21c1b0a115102540c204e