Extract fun types instead of coercing
Summary:
`Narrow.extractFunTypes` is mostly identical to `Narrow.asFunType` except it filters out non compatible types, while the latter performs an early return with `None` in that case.

However, `Narrow.asFunType` is always used with `.get` and after a check that the type is actually compatible, thus defeating the point of the optional result.

This diff migrates everything to `extractFunTypes`, the semantics are unchanged.

Reviewed By: ilya-klyuchnikov

Differential Revision: D62570404

fbshipit-source-id: 9ee92cb7bfe736df56032f41fcc89ef05bae629e