Typed expressions
Summary:
Implementation of `TypedExpr` on eqWAlizer's end.

`TypedExpr` wraps non-deserialized JSON data corresponding to a non-expanded, non-validated `ExtType` on ELP's side.

During type-checking, the json value is sent to ELP for validation, which sends back either:
1. A validated `Type`
2. An `Invalid` if the expansion failed
3. An `InvalidRefInTypeAnno` if the validation failed because the annotation references an invalid type
4. A `VariablesInTypeAnno` if the type annotation references type variables
5. A `RefinedRecordInTypeAnno` if the type annotation contains a refined record type

This needs a uniformisation of diagnostic types to use the super-trait `Diagnostic` instead of `Invalid` or `TypeError`.

Reviewed By: ilya-klyuchnikov

Differential Revision: D71735298

fbshipit-source-id: bbf8935a0f99519fede20bf802e26a6c7e30eb1d