Fix occurrence typing for is_function/2
Summary:
The actual top type for a function of arity `n` is `FunType(Nil, List.fill(n)(NoneType), AnyType)`, not `FunType(Nil, List.fill(n)(AnyType), AnyType)`, which means the negation of the predicate is not properly computed.

However, due to how intersection is computed (and particularly when it involves dynamic types), using this as a top type for the positive refinement leads to confusing signal, producing `none()` as domain. So this diff only uses it for the negative part.

Reviewed By: michalmuskala

Differential Revision: D85336038

fbshipit-source-id: 663393fe5919274c945435e764132faa4b0e7221